// SDKs/amazonaws.com/eventbridge/2015-10-07/typescript/src/sdk/models/shared/describeapidestinationresponse.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";
import { ApiDestinationStateEnum } from "./apidestinationstateenum";
import { ApiDestinationHttpMethodEnum } from "./apidestinationhttpmethodenum";


export class DescribeApiDestinationResponse extends SpeakeasyBase {
  @Metadata({ data: "json, name=ApiDestinationArn" })
  apiDestinationArn?: string;

  @Metadata({ data: "json, name=ApiDestinationState" })
  apiDestinationState?: ApiDestinationStateEnum;

  @Metadata({ data: "json, name=ConnectionArn" })
  connectionArn?: string;

  @Metadata({ data: "json, name=CreationTime" })
  creationTime?: Date;

  @Metadata({ data: "json, name=Description" })
  description?: string;

  @Metadata({ data: "json, name=HttpMethod" })
  httpMethod?: ApiDestinationHttpMethodEnum;

  @Metadata({ data: "json, name=InvocationEndpoint" })
  invocationEndpoint?: string;

  @Metadata({ data: "json, name=InvocationRateLimitPerSecond" })
  invocationRateLimitPerSecond?: number;

  @Metadata({ data: "json, name=LastModifiedTime" })
  lastModifiedTime?: Date;

  @Metadata({ data: "json, name=Name" })
  name?: string;
}